requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Computer Vision
opencv-python==4.8.1.78
ultralytics==8.0.206
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.1.0

# Face Recognition
face-recognition==1.3.0
dlib==19.24.2

# Database and Storage
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.12.1

# HTTP Client and API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Configuration and Logging
pyyaml==6.0.1
python-dotenv==1.0.0
structlog==23.2.0

# Additional Production Dependencies
psutil>=5.9.0
scipy>=1.10.0
matplotlib>=3.7.0
websockets>=12.0
structlog==23.2.0
colorama==0.4.6

# Async and Concurrency
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Raspberry Pi Specific
RPi.GPIO==0.7.1
picamera2==0.3.12

# System Monitoring
psutil==5.9.6
GPUtil==1.4.0

# Date and Time
python-dateutil==2.8.2
pytz==2023.3
